// stock_react_miniproject/src/component/utils/Interfaces.ts
// Generated by https://quicktype.io

export interface IStockAPI {
  stock_name: string;
  stock_symbol: string;
  base_price: number;
  stock_status?: string;
}

export interface IStockSummary {
  company: string;
  symbol: string;
  data: Datum[];
}

export interface Datum {
  date: string;
  prices: number[];
}

export interface IStockTransaction {
  stock_name: string;
  stock_symbol: string;
  transaction_price: number;
  timestamp: string;
  status: string;
}